Gillnet Closure Areas

An sf object containing the Closure areas for Gillnets Specifically Closed Area 1, and Nantucket Lightship Closure Area

GillnetClosureAreas_20211021

Format

A data frame/ sf object (2 x 25)

GARFO_ID

A unique identifier used to identify and track this feature; a new GARFO_ID is generated for each version of a Regulated Area, including proposed or final versions (i.e. whenever boundaries or attributes are changed, edited or updated)

AREANAME

Official name of the Proposed Area, usually the area name as it will be printed in the CFR.

COMMNAME

Most commonly used name. May be identical to AREANAME, an abbreviation of AREANAME, or a different name altogether.

AREAGROUP

Specifies if the Proposed Area is part of a group of related areas.

DESCRIBE

A brief description of the purpose of the Proposed Area.

CFRTITLE

CFR Title citation where the Proposed Area will be legally defined

CFRPT

ACFR Part citation where the Proposed Area will be legally defined

CFRPTTXT

CFR Part citation header text

CFRSUB

CFR Subpart citation where the Proposed Area will be legally defined

CFRSUBTXT

CFR Subpart citation header text

CFRSXN

CFR Section citation where the Proposed Area will be legally defined

CFRSXNTXT

CFR Section citation header text

CFRPARA

CFR Paragraph citation where the Proposed Area will be legally defined

CFRPARATXT

CFR Paragraph citation header text

FRCITE

Federal Register citation of proposed rule containing the Proposed Area spatial definition.

FRDATE

Federal Register citation date of proposed rule containing the Proposed Area spatial definition.

EFFECTDATE

Date Regulated Area officially became law, or most recent modification. For Proposed Areas, this field will be blank because the area is not yet law.

SOURCE

If Proposed Area will be defined in the CFR, this is the citation for the Amendment/Framework/etc. that will establish the area. If the Proposed Area will not be defined in the CFR, this is the alternative source citation that will establish the area.

RECURST

Annually recurring date Proposed Area would become active

RECUREND

Annually recurring date Proposed Area would become inactive

GISAUTHOR

Name of the NOAA staff person who created the GIS feature

CREATED

Creation date of the GIS feature

AREANOTES

Miscellaneous notes regarding the Proposed Area or feature

X

x coordinate of polygon center

Y

y coordinate of polygon center

Source

https://www.fisheries.noaa.gov/resource/map/gillnet-closure-areas

Contact

andrew.beet at noaa.gov